Finding the perfect spot to open a Vegetarian high-end Restaurant in Paris
Welcome to the capstone project of Alessandro C. The purpose of this project is to find the perfect arrondissement (the french word for "neighbourhood") to open an high-end vegetarian restaurant in Paris, France. This project uses Python and many of its Data Science libraries including pandas, numpy, sklearn, matplotlib, requests, beautifulsoup and seaborn.